requirements.txt
Django==4.0.4
djangorestframework==3.14.0
psycopg2==2.9.3
python-decouple==3.4